ml_variants/pytorch/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py39h7deecbd_0
brotlipy=0.7.0=py39h27cfd23_1003
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
conda=22.11.1=py39hf3d152e_1
conda-content-trust=0.1.1=pyhd3eb1b0_0
conda-package-handling=1.8.1=py39h7f8727e_0
cryptography=36.0.0=py39h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
flit-core=3.6.0=pyhd3eb1b0_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pyhd8ed1ab_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=12_linux64_mkl
libllvm11=11.1.0=h9e868ea_6
libstdcxx-ng=11.2.0=h1234567_1
libuv=1.40.0=h7b6447c_0
llvmlite=0.38.0=py39h4ff587b_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h7f8727e_2
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numba=0.55.1=py39h51133e4_0
numexpr=2.8.4=py39he184ba9_0
numpy=1.21.5=py39h6c91a56_3
numpy-base=1.21.5=py39ha15fc14_3
openssl=1.1.1s=h7f8727e_0
packaging=22.0=pyhd8ed1ab_0
pandas=1.5.2=py39h417a72b_0
pip=21.2.4=py39h06a4308_0
pluggy=1.0.0=py39h06a4308_1
pycosat=0.6.3=py39h27cfd23_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.8.2=py3.9_cuda10.2_cudnn7.6.5_0
pytz=2022.6=pyhd8ed1ab_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
ruamel.yaml=0.17.21=py39h5eee18b_0
ruamel.yaml.clib=0.2.6=py39h5eee18b_1
ruamel_yaml=0.15.100=py39h27cfd23_0
scikit-learn=0.24.2=py39h7c5d8c9_1
scipy=1.8.1=py39he49c0e8_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=hc218d9a_0
tbb=2021.6.0=hdb19cb5_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.11=h1ccaba5_0
toolz=0.12.0=py39h06a4308_0
tqdm=4.63.0=pyhd3eb1b0_0
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.12=h7f8727e_1